Parallels is only supported on darwin
